refactor: Rule names validation (#3406)
* refactor: UserInfo validation


* refactor: Rule type validation


* refactor: Rule names validation
