+++ api/kyverno/v1/spec_types.go
@@ -1,0 +1,141 @@
+package v1
+
+import (
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/apimachinery/pkg/util/validation/field"
+)
+
+// Spec contains a list of Rule instances and other policy controls.
+type Spec struct {
+	// Rules is a list of Rule instances. A Policy contains multiple rules and
+	// each rule can validate, mutate, or generate resources.
+	Rules []Rule `json:"rules,omitempty" yaml:"rules,omitempty"`
+
+	// FailurePolicy defines how unrecognized errors from the admission endpoint are handled.
+	// Rules within the same policy share the same failure behavior.
+	// Allowed values are Ignore or Fail. Defaults to Fail.
+	// +optional
+	FailurePolicy *FailurePolicyType `json:"failurePolicy,omitempty" yaml:"failurePolicy,omitempty"`
+
+	// ValidationFailureAction controls if a validation policy rule failure should disallow
+	// the admission review request (enforce), or allow (audit) the admission review request
+	// and report an error in a policy report. Optional. The default value is "audit".
+	// +optional
+	// +kubebuilder:validation:Enum=audit;enforce
+	ValidationFailureAction string `json:"validationFailureAction,omitempty" yaml:"validationFailureAction,omitempty"`
+
+	// ValidationFailureActionOverrides is a Cluster Policy attribute that specifies ValidationFailureAction
+	// namespace-wise. It overrides ValidationFailureAction for the specified namespaces.
+	// +optional
+	ValidationFailureActionOverrides []ValidationFailureActionOverride `json:"validationFailureActionOverrides,omitempty" yaml:"validationFailureActionOverrides,omitempty"`
+
+	// Background controls if rules are applied to existing resources during a background scan.
+	// Optional. Default value is "true". The value must be set to "false" if the policy rule
+	// uses variables that are only available in the admission review request (e.g. user name).
+	// +optional
+	Background *bool `json:"background,omitempty" yaml:"background,omitempty"`
+
+	// SchemaValidation skips policy validation checks.
+	// Optional. The default value is set to "true", it must be set to "false" to disable the validation checks.
+	// +optional
+	SchemaValidation *bool `json:"schemaValidation,omitempty" yaml:"schemaValidation,omitempty"`
+
+	// WebhookTimeoutSeconds specifies the maximum time in seconds allowed to apply this policy.
+	// After the configured time expires, the admission request may fail, or may simply ignore the policy results,
+	// based on the failure policy. The default timeout is 10s, the value must be between 1 and 30 seconds.
+	WebhookTimeoutSeconds *int32 `json:"webhookTimeoutSeconds,omitempty" yaml:"webhookTimeoutSeconds,omitempty"`
+}
+
+func (s *Spec) GetRules() []Rule {
+	return s.Rules
+}
+
+func (s *Spec) SetRules(rules []Rule) {
+	s.Rules = rules
+}
+
+// HasMutateOrValidateOrGenerate checks for rule types
+func (s *Spec) HasMutateOrValidateOrGenerate() bool {
+	for _, rule := range s.Rules {
+		if rule.HasMutate() || rule.HasValidate() || rule.HasGenerate() {
+			return true
+		}
+	}
+	return false
+}
+
+// HasMutate checks for mutate rule types
+func (s *Spec) HasMutate() bool {
+	for _, rule := range s.Rules {
+		if rule.HasMutate() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// HasValidate checks for validate rule types
+func (s *Spec) HasValidate() bool {
+	for _, rule := range s.Rules {
+		if rule.HasValidate() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// HasGenerate checks for generate rule types
+func (s *Spec) HasGenerate() bool {
+	for _, rule := range s.Rules {
+		if rule.HasGenerate() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// HasVerifyImages checks for image verification rule types
+func (s *Spec) HasVerifyImages() bool {
+	for _, rule := range s.Rules {
+		if rule.HasVerifyImages() {
+			return true
+		}
+	}
+
+	return false
+}
+
+// BackgroundProcessingEnabled checks if background is set to true
+func (s *Spec) BackgroundProcessingEnabled() bool {
+	if s.Background == nil {
+		return true
+	}
+
+	return *s.Background
+}
+
+// ValidateRuleNames checks if the rule names are unique across a policy
+func (s *Spec) ValidateRuleNames(path *field.Path) field.ErrorList {
+	var errs field.ErrorList
+	names := sets.NewString()
+	for i, rule := range s.Rules {
+		rulePath := path.Index(i)
+		if names.Has(rule.Name) {
+			errs = append(errs, field.Invalid(rulePath.Child("name"), rule, fmt.Sprintf(`Duplicate rule name: '%s'`, rule.Name)))
+		}
+		names.Insert(rule.Name)
+	}
+	return errs
+}
+
+// Validate implements programmatic validation
+func (s *Spec) Validate(path *field.Path) field.ErrorList {
+	var errs field.ErrorList
+	errs = append(errs, s.ValidateRuleNames(path.Child("rules"))...)
+	return errs
+}
